+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Remove CircleCI setup
 * Cleanup bookingsync_portal config generator and dummy app one
+* Add missing dummy simple_form initializer
 
 # 0.6.0