+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # master
 
 * Remove CircleCI setup
+* Cleanup bookingsync_portal config generator and dummy app one
+* Add missing dummy simple_form initializer
 
 # 0.6.0